requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
django==5.0; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-filter==23.5; python_version >= '3.7'
djangorestframework==3.14.0; python_version >= '3.6'
pytz==2023.3.post1
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.9.0; python_version < '3.11'
